services/appointment-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.3.0
dj-database-url==2.1.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
redis==5.0.1
celery==5.3.4
requests==2.31.0
django-filter==23.3
# Sẽ được cài đặt trong Dockerfile
# common-auth
